+++ tests/IntegrationTests/DI/PreConstructorInjectionTest.php
@@ -1,0 +1,33 @@
+<?php
+/**
+ * PHP-DI
+ *
+ * @link      http://mnapoli.github.com/PHP-DI/
+ * @copyright 2012 Matthieu Napoli (http://mnapoli.fr/)
+ * @license   http://www.opensource.org/licenses/mit-license.php MIT (see the LICENSE file)
+ */
+
+namespace IntegrationTests\DI;
+
+use DI\Container;
+use IntegrationTests\DI\Fixtures\PreConstructorInjection\Class1;
+
+/**
+ * Testing the injection is done before the constructor is called
+ */
+class PreConstructorInjectionInjectionTest extends \PHPUnit_Framework_TestCase
+{
+
+	public function setUp() {
+		// Reset the singleton instance to ensure all tests are independent
+		Container::reset();
+	}
+
+	public function testDependenciesAreInjectedBeforeConstructorIsCalled() {
+		$container = Container::getInstance();
+		/** @var $class1 Class1 */
+		$class1 = $container->get('IntegrationTests\DI\Fixtures\PreConstructorInjection\Class1');
+		$this->assertTrue($class1->dependencyInjected);
+	}
+
+}