+++ change-log.md
@@ -2,13 +2,19 @@
 
 ## 2.0
 
-Compatibility break with 1.x.
+Major compatibility breaks with 1.x.
 
-* Merge `@Value` annotation with `@Inject`: no difference between value and bean injection anymore
+* `Container::resolveDependencies()` has been renamed to `Container::injectAll()`
+* Dependencies are now injected **before** the constructor is called, and thus are available in the constructor
+* Merged `@Value` annotation with `@Inject`: no difference between value and bean injection anymore
 * Container implements ArrayAccess for get() and set() (`$container['db.host'] = 'localhost';`)
 * Ini configuration files removed: configuration is done in PHP
 * Allow to define beans within closures for lazy-loading
 * Switched to MIT License
+
+Warning:
+
+* If you use PHP 5.3 and __wakeup() methods, they will be called when PHP-DI creates new instances of them.
 
 ## 1.1